Move settings to separate module
